% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDouhauDataStudio.R
\name{runDouhauDataStudio}
\alias{runDouhauDataStudio}
\title{Douhau Data Studio Launcher}
\usage{
runDouhauDataStudio(
  dataDir,
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\arguments{
\item{dataDir}{A path sting of a data directory}

\item{port}{A port number as integer value should listen on. If this argument
is not specified, the shiny.port option is configured automatically.}

\item{launch.browser}{launch.browser argument can accept a logical value. If
FALSE is populated for this argument, the default browser will not
launched. And in case of TRUE, default will be launted. If this argument is
not specified, the shiny.port option is configured automatically.}
}
\description{
Douhau Data Studio Launcher
}
