% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDouhauDataStudio.R
\name{runDouhauDataStudio}
\alias{runDouhauDataStudio}
\title{Douhau Data Studio Launcher}
\usage{
runDouhauDataStudio(dataDir)
}
\arguments{
\item{dataDir}{A path sting of a data directory}
}
\description{
Douhau Data Studio Launcher
}
